requirements.txt
fastapi==0.100.1
uvicorn==0.23.2
SQLAlchemy==2.0.19
python-dotenv==1.0.0
pydantic==2.1.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
pytest==7.4.0
pytest-asyncio==0.21.1
httpx==0.24.1
PyJWT==2.8.0
psycopg2-binary==2.9.6
python-creditcard==1.0.0